requirements.txt
requests~=2.31.0
Flask~=2.0.3
summarizer~=0.0.7
g4f~=0.3.1.9
openai~=1.30.5
protobuf==3.19.4
transformers
azure-storage-queue
azure-functions==1.19.0
bert-extractive-summarizer==0.10.1
boto3==1.23.10
docopt==0.6.2
Flask==2.0.3
install==1.3.5
Pillow==8.4.0

pip==21.3.1
pytest==7.0.1
sentencepiece==0.2.0
summarizer==0.0.7
tensorflow==2.8.4
torch
yarg==0.1.9
opencv-python~=4.9.0.80
azure-mgmt-compute
azure-mgmt-storage
azure-mgmt-resource
azure-storage-blob
azure-data-tables
